+++ Cargo.toml
@@ -4,8 +4,6 @@ version = "0.1.0"
 authors = ["iquerejeta <querejeta.azurmendi@iohk.io>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 rand_core = { version = "0.5", default-features = false }
 generic-array = "0.14"
