+++ src/dkg/mod.rs
@@ -1,3 +1,4 @@
 pub mod broadcast;
 pub mod committee;
+pub mod fetched_state;
 pub mod procedure_keys;
