+++ src/errors.rs
@@ -76,5 +76,3 @@ impl From<ProofError> for DkgError {
         DkgError::ZkpVerificationFailed
     }
 }
-
-// todo: we probably want broadcast errors, to capture misbehaving parties
